+++ packages/react-native-eyecandy/ios/ReactNativeEyecandy.swift
@@ -5,4 +5,9 @@ class ReactNativeEyecandy: NSObject {
     func multiply(a: Float, b: Float, resolve:RCTPromiseResolveBlock,reject:RCTPromiseRejectBlock) -> Void {
         resolve(a*b)
     }
+    
+    @objc
+    static func requiresMainQueueSetup() -> Bool {
+        return true;
+    }
 }