+++ packages/react-native-eyecandy/src/components/ButtonBase/useStyles.ts
@@ -14,7 +14,7 @@ export default function useStyles({
   variant = 'default',
   height = 56,
   disabled,
-  fullwidth = true,
+  fullwidth = false,
   transparent = false,
   outlined = false,
 }: {
@@ -57,12 +57,15 @@ export default function useStyles({
       borderWidth,
       borderColor,
       backgroundColor: backgroundColor.string(),
+      flex: 1,
     },
     pressable: {
       flex: 1,
       borderRadius,
+      flexDirection: 'row',
       justifyContent: 'center',
       alignItems: 'center',
+      paddingHorizontal: 16,
     },
     ripple: {
       color: rippleColor,