+++ packages/react-native-eyecandy/src/components/DatePicker/Calendar/Month/index.ts
@@ -1,4 +1,0 @@
-import Month from './Month';
-
-export * from './Month';
-export default Month;