+++ packages/react-native-eyecandy/src/components/DatePicker/Calendar/index.ts
@@ -1,4 +1,0 @@
-import Calendar from './Calendar';
-
-export * from './Calendar';
-export default Calendar;