+++ packages/react-native-eyecandy/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nomada-sh/react-native-eyecandy
 
+## 0.0.6
+
+### Patch Changes
+
+- Created ProgressButton component.
+- Added `dragOutEnabled` prop to BottomSheet component.
+- Added `@react-native-masked-view/masked-view@0.2.x` as peerDependency.
+- Updated dependencies
+  - @nomada-sh/react-native-eyecandy-theme@0.0.3
+  - @nomada-sh/react-native-eyecandy-icons@0.0.5
+
 ## 0.0.5
 
 ### Patch Changes
@@ -14,20 +25,20 @@
 
 ### Patch Changes
 
-- Renamed component Passcode to Dialer
+- Renamed component Passcode to Dialer.
 
 ## 0.0.3
 
 ### Patch Changes
 
-- Improved Select component UX
+- Improved Select component UX.
 
 ## 0.0.2
 
 ### Patch Changes
 
-- Removed NativeSelect component
-- Changed peerDependencies versions
+- Removed NativeSelect component.
+- Changed peerDependencies versions.
 - Updated dependencies
   - @nomada-sh/react-native-eyecandy-icons@0.0.3
   - @nomada-sh/react-native-eyecandy-theme@0.0.2